Rework well under way. Tether spring for each note plated. Labeled sliders. Needs octave spring, etc.

+++ bk_JUCE/bitKlavier/Source/SpringTuningUtilities.h
@@ -109,7 +109,7 @@ public:
 	{
 		return (int)(ratioToCents(whichFreq / cFreq));
 	}
-	static double centsToFreq(int centsFromC)
+	static double centsToFreq(double centsFromC)
 	{
 		return cFreq * pow(2, centsFromC / 1200.0);
 	}
